Fixed a warning in saveOggOpus

+++ libAudio/saveOggOpus.cxx
@@ -113,7 +113,7 @@ int64_t oggOpus_t::writeBuffer(const void *const bufferPtr, const int64_t length
 	else if (info.bitsPerSample != 16)
 		return -3; // Opus can't encode non-16-bit sample data.. V_V
 	// Convert length into samples per channel
-	const auto sampleCount{uint64_t(length) / (info.channels * (info.bitsPerSample / 8U))};
+	const auto sampleCount{uint64_t(length) / (uint64_t(info.channels) * (info.bitsPerSample / 8U))};
 	const auto *const buffer{static_cast<const int16_t *>(bufferPtr)};
 	const auto result{ope_encoder_write(ctx.encoder, buffer, sampleCount)};
 	if (result == OPE_OK)
